Second version of the LSH algorithm.
This code fetches the education of the scientists in range, creates the set of shingles for each education (scientist) and the universal set of shingles and creates the input matrix for the MinHashing part of the LSH algorithm. It also executes the MinHashing part of the algorithm and creates each document's signature.